refactor(graduate): 로그아웃 및 라우터 분리
