+++ apps/graduate/src/routes/__root.tsx
@@ -1,13 +1,12 @@
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { Outlet, createRootRouteWithContext } from '@tanstack/react-router';
 import { TanStackRouterDevtools } from '@tanstack/react-router-devtools';
+
+import { useAuthStore } from '~/shared/stores';
 
 import { Header } from '~/widgets/Header';
 import { Sidebar } from '~/widgets/sidebar';
 
 import { vars } from '~/vars.css';
-
-const queryClient = new QueryClient();
 
 export interface AuthContext {
   auth: {
@@ -20,16 +19,16 @@ export interface AuthContext {
 
 export const Route = createRootRouteWithContext<AuthContext>()({
   component: () => {
-    const { auth } = Route.useRouteContext();
+    const { isAuthenticated, isAdmin } = useAuthStore();
 
-    if (!auth.isAuthenticated) {
+    if (!isAuthenticated) {
       return <Outlet />;
     }
 
-    if (auth.isAuthenticated && auth.isAdmin) {
+    if (isAuthenticated && isAdmin) {
       return (
         <div style={{ display: 'flex' }}>
-          <Sidebar auth={auth} />
+          <Sidebar />
           <main
             style={{
               flex: '1 1 auto',
@@ -43,10 +42,10 @@ export const Route = createRootRouteWithContext<AuthContext>()({
       );
     }
 
-    if (auth.isAuthenticated && !auth.isAdmin) {
+    if (isAuthenticated && !isAdmin) {
       return (
-        <QueryClientProvider client={queryClient}>
-          <body
+        <>
+          <div
             style={{
               position: 'relative',
               minHeight: '100dvh',
@@ -54,7 +53,7 @@ export const Route = createRootRouteWithContext<AuthContext>()({
               backgroundColor: vars.colors.sub,
             }}
           >
-            <Header auth={auth} />
+            <Header />
             <main
               style={{
                 position: 'absolute',
@@ -68,9 +67,9 @@ export const Route = createRootRouteWithContext<AuthContext>()({
             >
               <Outlet />
             </main>
-          </body>
+          </div>
           <TanStackRouterDevtools />
-        </QueryClientProvider>
+        </>
       );
     }
   },
