+++ apps/graduate/src/shared/hooks/index.ts
@@ -1,2 +1,1 @@
 export { default as useTableState } from './useTableState';
-export { default as useLogout } from './useLogout';
