+++ apps/graduate/src/widgets/Header/ui/Header.tsx
@@ -2,20 +2,16 @@ import { Link } from '@tanstack/react-router';
 import { Avatar, Dropdown, MenuProps } from 'antd';
 import { User } from 'lucide-react';
 
-import { useLogout } from '~/shared/hooks';
+import { logout } from '~/shared/utils';
 
 import * as styles from '../styles/Header.css';
 
-import { AuthContext } from '~/routes/__root';
-
-export default function Header({ auth }: AuthContext) {
-  const handleLogout = useLogout(auth);
-
+export default function Header() {
   const items: MenuProps['items'] = [
     {
       key: '1',
       label: (
-        <a target='_blank' rel='noopener noreferrer' onClick={handleLogout}>
+        <a target='_blank' rel='noopener noreferrer' onClick={logout}>
           로그아웃
         </a>
       ),
